# api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "hivent-groups-api"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "groups"
- name: "workshops"
- name: "communities"
- name: "projects"
paths:
  /groups:
    get:
      tags:
      - "groups"
      operationId: "findGroups"
      parameters:
      - name: "name"
        in: "query"
        required: false
        type: "string"
      - name: "status"
        in: "query"
        description: "Status [ PROPOSED, OPENED, CLOSED ]"
        required: false
        type: "string"
        enum:
        - "PROPOSED"
        - "OPENED"
        - "CLOSED"
      - name: "type"
        in: "query"
        description: "Type [ PROJECT, WORKSHOP, COMMUNITY ]"
        required: false
        type: "string"
        enum:
        - "PROJECT"
        - "WORKSHOP"
        - "COMMUNITY"
      - name: "userId"
        in: "query"
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Groups"
  /workshops:
    get:
      tags:
      - "workshops"
      operationId: "findWorkshops"
      parameters:
      - name: "name"
        in: "query"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "csv"
      - name: "selectionMode"
        in: "query"
        description: "SelectionMode [ NOLIMIT, RANDOM, FIRST ]"
        required: false
        type: "array"
        items:
          type: "string"
          enum:
          - "NOLIMIT"
          - "RANDOM"
          - "FIRST"
        collectionFormat: "csv"
      - name: "status"
        in: "query"
        description: "Status [ PROPOSED, OPENED, CLOSED ]"
        required: false
        type: "array"
        items:
          type: "string"
          enum:
          - "PROPOSED"
          - "OPENED"
          - "CLOSED"
        collectionFormat: "csv"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
    post:
      tags:
      - "workshops"
      operationId: "createWorkshop"
      parameters:
      - in: "body"
        name: "workshop"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
  /workshops/{id}:
    get:
      tags:
      - "workshops"
      operationId: "readWorkshop"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
    put:
      tags:
      - "workshops"
      operationId: "updateWorkshop"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "Workshop"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        200:
          description: "Mise à jour réussie."
          schema:
            $ref: "#/definitions/Group"
        400:
          description: "Erreur lors de la mise à jour."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
    delete:
      tags:
      - "workshops"
      operationId: "deleteWorkshop"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
  /workshops/{id}/administrators:
    post:
      tags:
      - "workshops"
      operationId: "createWorkshopAdministrators"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        description: "id seulement"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
  /workshops/{id}/interested:
    post:
      tags:
      - "workshops"
      operationId: "createWorkshopInterested"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        description: "id seulement"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Workshops"
  /communities:
    get:
      tags:
      - "communities"
      operationId: "findCommunities"
      parameters:
      - name: "name"
        in: "query"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "csv"
      - name: "status"
        in: "query"
        description: "Status [ PROPOSED, OPENED, CLOSED ]"
        required: false
        type: "array"
        items:
          type: "string"
          enum:
          - "PROPOSED"
          - "OPENED"
          - "CLOSED"
        collectionFormat: "csv"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
            minItems: 1
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
    post:
      tags:
      - "communities"
      operationId: "createCommunity"
      parameters:
      - in: "body"
        name: "community"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
  /communities/{id}:
    get:
      tags:
      - "communities"
      operationId: "readCommunity"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
    put:
      tags:
      - "communities"
      operationId: "updateCommunity"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "community"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        200:
          description: "Mise à jour réussie."
          schema:
            $ref: "#/definitions/Group"
        400:
          description: "Erreur lors de la mise à jour."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
    delete:
      tags:
      - "communities"
      operationId: "deleteCommunity"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
  /communities/{id}/members:
    post:
      tags:
      - "communities"
      operationId: "createCommunityMembers"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        description: "id de l'utilisateur"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
  /communities/{id}/administrators:
    post:
      tags:
      - "communities"
      operationId: "createCommunityAdministrators"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
  /communities/{id}/interested:
    post:
      tags:
      - "communities"
      operationId: "createCommunityInterested"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Communities"
  /projects:
    get:
      tags:
      - "projects"
      operationId: "findProjects"
      parameters:
      - name: "status"
        in: "query"
        description: "Status [ PROPOSED, OPENED, CLOSED ]"
        required: false
        type: "array"
        items:
          type: "string"
          enum:
          - "PROPOSED"
          - "OPENED"
          - "CLOSED"
        collectionFormat: "csv"
      - name: "name"
        in: "query"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "csv"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
    post:
      tags:
      - "projects"
      operationId: "createProject"
      parameters:
      - in: "body"
        name: "project"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
  /projects/{id}:
    get:
      tags:
      - "projects"
      operationId: "readProject"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Lecture réussie."
          schema:
            $ref: "#/definitions/Group"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
    put:
      tags:
      - "projects"
      operationId: "updateProject"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "project"
        required: true
        schema:
          $ref: "#/definitions/Group"
      responses:
        200:
          description: "Mise à jour réussie."
          schema:
            $ref: "#/definitions/Group"
        400:
          description: "Erreur lors de la mise à jour."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
    delete:
      tags:
      - "projects"
      operationId: "deleteProject"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Suppression réussie."
        400:
          description: "Erreur lors de la suppression."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
  /projects/{id}/members:
    post:
      tags:
      - "projects"
      operationId: "createProjectMembers"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
  /projects/{id}/administrators:
    post:
      tags:
      - "projects"
      operationId: "createProjectAdministrators"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
  /projects/{id}/interested:
    post:
      tags:
      - "projects"
      operationId: "createProjectInterested"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "users"
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          minItems: 1
      responses:
        201:
          description: "Création réussie."
          schema:
            $ref: "#/definitions/Group"
          headers:
            Location:
              type: "string"
              description: "URI de la ressource créée"
        400:
          description: "Erreur lors de la création."
          schema:
            $ref: "#/definitions/ReponseErreur"
        404:
          description: "Ressource introuvable."
          schema:
            $ref: "#/definitions/ReponseErreur"
      x-swagger-router-controller: "Projects"
definitions:
  Group:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Identifiant.Lecture seule.\n"
      name:
        type: "string"
      description:
        type: "string"
      proposedDate:
        type: "string"
        format: "date"
      openedDate:
        type: "string"
        format: "date"
      closedDate:
        type: "string"
        format: "date"
      capacity:
        type: "integer"
        format: "int32"
      pathImage:
        type: "string"
      administrators:
        type: "array"
        items:
          $ref: "#/definitions/User"
        minItems: 1
      interested:
        type: "array"
        items:
          $ref: "#/definitions/User"
      members:
        type: "array"
        items:
          $ref: "#/definitions/User"
      type:
        $ref: "#/definitions/Type"
      status:
        $ref: "#/definitions/Status"
      defaultEventSelectionMode:
        $ref: "#/definitions/SelectionMode"
    description: "Communities and Projects"
    example:
      description: "description"
      type: {}
      capacity: 6
      openedDate: "2000-01-23"
      pathImage: "pathImage"
      closedDate: "2000-01-23"
      proposedDate: "2000-01-23"
      members:
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      name: "name"
      defaultEventSelectionMode: {}
      id: 0
      interested:
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      administrators:
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      - firstname: "firstname"
        pathImage: "pathImage"
        mail: "mail"
        id: 1
        lastname: "lastname"
      status: {}
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Identifiant.Lecture seule.\n"
      firstname:
        type: "string"
      lastname:
        type: "string"
      mail:
        type: "string"
      pathImage:
        type: "string"
    example:
      firstname: "firstname"
      pathImage: "pathImage"
      mail: "mail"
      id: 1
      lastname: "lastname"
  Status:
    description: "Status"
    enum:
    - "PROPOSED"
    - "OPENED"
    - "CLOSED"
  Type:
    description: "Type"
    enum:
    - "PROJECT"
    - "WORKSHOP"
    - "COMMUNITY"
  SelectionMode:
    description: "SelectionMode"
    enum:
    - "NOLIMIT"
    - "RANDOM"
    - "FIRST"
  ReponseErreur:
    type: "object"
    properties:
      erreurs:
        type: "array"
        items:
          $ref: "#/definitions/ReponseErreur_erreurs"
  ReponseErreur_erreurs:
    properties:
      champErreur:
        type: "string"
      codeErreur:
        type: "string"
      libelleErreur:
        type: "string"